// src/types.ts
export interface LeadPayload {
  fullname: string;
  cpf: string;
  whatsapp: string;
  email: string;
  profession: string;
  country: string;
  cep: string;
  street: string;
  number: string;
  complement: string;
  neighborhood: string;
  city: string;
  state: string;
  seeking_reason: string;
  attention_reason: string;
  createdAt: any;
}